C++: switch to size_t in uArray and uString
size_t is commonly used to represent indices and lengths in C/C++ code, and
instead of adding casts to int32_t, we change the interface of uArray and
uString to accept size_t. This seems to align better with a lot of C/C++
code, and is non-breaking since int32_t implicitly converts to size_t and
vice versa. Since many call sites were size_t to begin with we can remove
some explicit casts to int32_t, and since we can remove some often called
"< 0" checks the code is in theory a bit faster now than it was before.

This fixes a few instances of the following warning:

    warning C4267: 'argument': conversion from 'size_t' to 'int32_t', possible loss of data

+++ lib/UnoCore/Backends/CPlusPlus/Uno/Memory.cpp
@@ -723,7 +723,7 @@ uObject* uNew(uType* type, size_t size)
     return uInitObject(type, calloc(1, size), size);
 }
 
-static uString* uInitString(int32_t length)
+static uString* uInitString(size_t length)
 {
     size_t size = sizeof(uString) + sizeof(char16_t) * length + sizeof(char16_t);
     uString* string = (uString*)uInitObject(@{string:TypeOf}, calloc(1, size), size);
@@ -732,7 +732,7 @@ static uString* uInitString(int32_t length)
     return string;
 }
 
-uString* uString::New(int32_t length)
+uString* uString::New(size_t length)
 {
     if (length == 0)
     {
